Report server address when using an AF_UNIX socket.
When uvicorn is listening on an AF_UNIX socket (--uds), the "server"
scope parameter should be (socket_path, None).  This allows the
application/middleware to detect that a Unix socket is in use.

The test_get_local_addr test case needs to be updated accordingly;
previously it was assumed that get_local_addr would return None for a
Unix socket.

The test_get_local_addr_with_socket test case was nonsensical: if the
'family' is AF_UNIX then the 'sockname' must be a string.  Fix this
test case to check that get_local_addr works in both cases.