Add test cases for abstract Unix sockets.
On Linux, if an AF_UNIX socket is bound to an abstract
address (starting with a NUL byte), then Python's socket API reports
the address as a 'bytes' object rather than a 'str'.

Check that this is handled correctly by get_local_addr.  The server
address reported to the ASGI application should always be a Unicode
string.