Allow listing AF_UNIX clients as trusted for proxy headers.
When uvicorn is behind a reverse proxy, we want a way to report the
real client IP address and protocol based on the X-Forwarded-For and
X-Forwarded-Proto headers.  If we don't want people to be able to
impersonate arbitrary addresses, we need to identify which proxies are
trusted.

In order to allow AF_UNIX proxies to be trusted, use the string
"unix:" as a placeholder address.  Thus, "--forwarded-allow-ips=unix:"
means to trust AF_UNIX proxies but not IPv4 or IPv6 proxies.

(This syntax is borrowed from nginx: if an nginx proxy server is
itself listening on an AF_UNIX socket, "unix:" is what it will display
in X-Forwarded-For.  Thus, "--forwarded-allow-ips=unix:" allows us to
trust a chain of multiple AF_UNIX proxies.)